Python/Connect4+Tests/requirements.txt
pyhamcrest
parameterized
coverage
assertpy
nose